src/main/java/com/example/tobbyspring/basic/chpter5/step1/JdbcUserDao.java
package com.example.tobbyspring.basic.chpter5.step1;

import com.example.tobbyspring.entity.User;
import lombok.RequiredArgsConstructor;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;

import java.util.List;

@Component
@RequiredArgsConstructor
public class JdbcUserDao implements UserDao{

    private final JdbcTemplate jdbcTemplate;

    @Override
    public void add(User user) {
        this.jdbcTemplate.update("insert into users(name, password) values(?,?)", user.getName(), user.getPassword());
    }

    @Override
    public User get(Long id) {
        return this.jdbcTemplate.queryForObject("select * from users where id = ?", new Object[]{id}, this.userMapper);
    }

    @Override
    public List<User> getAll() {
        return this.jdbcTemplate.query("select * from users order by id", this.userMapper);
    }

    @Override
    public void deleteAll() {
        this.jdbcTemplate.update("delete from users");
    }

    @Override
    public int getCount() {
        return this.jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
    }

    private RowMapper<User> userMapper =
            (rs, rowNum) -> {
                User user = new User();
                user.setId(rs.getLong("id"));
                user.setName(rs.getString("name"));
                user.setPassword(rs.getString("password"));
                return user;
            };
}
